# packages/contracts/openapi/openapi.yaml
components:
  schemas:
...
    Device:
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        fcmToken:
          maxLength: 4096
          minLength: 1
          type: string
        id:
          minLength: 1
          type: string
        name:
          maxLength: 64
          minLength: 1
          type: string
        platform:
          enum:
            - windows
            - linux
            - android
        soundEnabled:
          type: boolean
      required:
        - id
        - name
        - platform
        - enabled
        - soundEnabled
      type: object
    DeviceListResponse:
      items:
        additionalProperties: false
        properties:
          enabled:
            type: boolean
          fcmToken:
            maxLength: 4096
            minLength: 1
            type: string
          id:
            minLength: 1
            type: string
          name:
            maxLength: 64
            minLength: 1
            type: string
          platform:
            enum:
              - windows
              - linux
              - android
          soundEnabled:
            type: boolean
        required:
          - id
          - name
          - platform
          - enabled
          - soundEnabled
        type: object
      type: array
    EmailBody:
      additionalProperties: false
      properties:
        email:
          format: email
          type: string
      required:
        - email
      type: object
    ErrorResponse:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              minLength: 1
              type: string
            message:
              minLength: 1
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
    LoginBody:
      additionalProperties: false
      properties:
        email:
          format: email
          type: string
        password:
          maxLength: 128
          minLength: 8
          type: string
      required:
        - email
        - password
      type: object
...
    PatchDeviceBody:
      additionalProperties: false
      minProperties: 1
      properties:
        enabled:
          type: boolean
        fcmToken:
          maxLength: 4096
          minLength: 1
          type: string
        name:
          maxLength: 64
          minLength: 1
          type: string
        soundEnabled:
          type: boolean
      type: object
    RefreshBody:
      additionalProperties: false
      properties:
        refreshToken:
          minLength: 1
          type: string
      required:
        - refreshToken
      type: object
    RegisterBody:
      additionalProperties: false
      properties:
        email:
          format: email
          type: string
        password:
          maxLength: 128
          minLength: 8
          type: string
      required:
        - email
        - password
      type: object
    RegisterDeviceBody:
      additionalProperties: false
      properties:
        enabled:
          default: true
          type: boolean
        fcmToken:
          maxLength: 4096
          minLength: 1
          type: string
        name:
          maxLength: 64
          minLength: 1
          type: string
        platform:
          enum:
            - windows
            - linux
            - android
        soundEnabled:
          default: true
          type: boolean
      required:
        - name
        - platform
      type: object
    ResetPasswordBody:
      additionalProperties: false
      properties:
        code:
          description: Eight-character high-entropy alphanumeric code delivered over SMTP.
          maxLength: 8
          minLength: 8
          pattern: ^[A-Za-z0-9]{8}$
          type: string
        email:
          format: email
          type: string
        password:
          maxLength: 128
          minLength: 8
          type: string
      required:
        - email
        - code
        - password
      type: object
    TokenPair:
      additionalProperties: false
      properties:
        accessToken:
          description: JWT access token, valid for 15 minutes.
          minLength: 1
          type: string
        refreshToken:
          description: Opaque refresh token, valid for 30 days; rotated on every refresh.
          minLength: 1
          type: string
      required:
        - accessToken
        - refreshToken
      type: object
    VerifyEmailBody:
      additionalProperties: false
      properties:
        code:
          description: Eight-character high-entropy alphanumeric code delivered over SMTP.
          maxLength: 8
          minLength: 8
          pattern: ^[A-Za-z0-9]{8}$
          type: string
        email:
          format: email
          type: string
      required:
        - email
        - code
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Short-lived (15 minute) JWT access token from login or refresh.
      scheme: bearer
      type: http
    ingestKeyAuth:
      description: Ingest credentials `keyId.secret`; every request must also carry valid X-Notify-Timestamp and X-Notify-Signature HMAC headers.
      scheme: bearer
      type: http
info:
  description: Single protocol source for the opencode-notify gateway, the OpenCode plugin, and the Flutter clients. Generated from the shared JSON schemas in @notify/contracts; do not edit by hand.
  license:
    identifier: MIT
    name: MIT
  title: opencode-notify gateway
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/auth/forgot-password:
    post:
      description: Returns the same response for known and unknown emails to avoid enumeration.
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailBody"
        required: true
      responses:
        "204":
          description: Reset email sent when the account exists.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
      security: []
      summary: Send a password reset email.
      tags:
        - auth
  /v1/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
          description: Session tokens.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid credentials.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Email address is not verified.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
      security: []
      summary: Log in with email and password.
      tags:
        - auth
  /v1/auth/logout:
    post:
      description: "Idempotent: unknown or already revoked tokens still return 204."
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
        required: true
      responses:
        "204":
          description: Refresh token revoked.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
      security: []
      summary: Revoke a refresh token.
      tags:
        - auth
  /v1/auth/refresh:
    post:
      description: Every refresh rotates the refresh token. Presenting an already rotated token revokes the whole token family (REFRESH_REUSED).
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
          description: Rotated session tokens.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid, expired, or reused refresh token.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
      security: []
      summary: Rotate a refresh token for a new token pair.
      tags:
        - auth
  /v1/auth/register:
    post:
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
        required: true
      responses:
        "201":
          description: Account created; verification email sent.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Email already registered.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Verification email delivery failed; the unverified account can resend later.
      security: []
      summary: Register an account and send a verification email.
      tags:
        - auth
  /v1/auth/resend-verification:
    post:
      description: Returns the same response for known and unknown emails to avoid enumeration.
      operationId: resendVerification
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailBody"
        required: true
      responses:
        "204":
          description: Verification email sent when the account exists.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Email delivery failed; retry later.
      security: []
      summary: Resend the verification email.
      tags:
        - auth
  /v1/auth/reset-password:
    post:
      description: A successful reset revokes every refresh-token family of the user.
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
        required: true
      responses:
        "204":
          description: Password reset.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid, expired, or already used code.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
      security: []
      summary: Reset the password with the SMTP-delivered code.
      tags:
        - auth
  /v1/auth/verify-email:
    post:
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailBody"
        required: true
      responses:
        "204":
          description: Email verified.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid, expired, or already used code.
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limited.
      security: []
      summary: Verify an email address with the SMTP-delivered code.
      tags:
        - auth
  /v1/devices:
    get:
      operationId: listDevices
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
          description: Registered devices.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
      security:
        - bearerAuth: []
      summary: List the authenticated user's devices.
      tags:
        - devices
    post:
      description: The server applies enabled/soundEnabled defaults when the body omits them.
      operationId: registerDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceBody"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
          description: Registered device.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
      security:
        - bearerAuth: []
      summary: Register a device.
      tags:
        - devices
  /v1/devices/{id}:
    delete:
      operationId: deleteDevice
      parameters:
        - description: Device identifier.
          in: path
          name: id
          required: true
          schema:
            minLength: 1
            type: string
      responses:
        "204":
          description: Device deleted.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unknown device for this user.
      security:
        - bearerAuth: []
      summary: Delete a device.
      tags:
        - devices
    patch:
      operationId: updateDevice
      parameters:
        - description: Device identifier.
          in: path
          name: id
          required: true
          schema:
            minLength: 1
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchDeviceBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
          description: Updated device.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unknown device for this user.
      security:
        - bearerAuth: []
      summary: Update a device.
      tags:
        - devices
...